+++ src/main/java/org/opentripplanner/api/mapping/PropertyMapper.java
@@ -1,10 +1,11 @@
-package org.opentripplanner.ext.vectortiles;
+package org.opentripplanner.api.mapping;
 
-import com.wdtinc.mapbox_vector_tile.VectorTile;
-import com.wdtinc.mapbox_vector_tile.adapt.jts.IUserDataConverter;
-import com.wdtinc.mapbox_vector_tile.adapt.jts.UserDataKeyValueMapConverter;
-import com.wdtinc.mapbox_vector_tile.build.MvtLayerProps;
+import edu.colorado.cires.cmg.mvt.VectorTile;
+import edu.colorado.cires.cmg.mvt.adapt.jts.IUserDataConverter;
+import edu.colorado.cires.cmg.mvt.adapt.jts.UserDataKeyValueMapConverter;
+import edu.colorado.cires.cmg.mvt.build.MvtLayerProps;
 import java.util.Collection;
+import org.opentripplanner.inspector.vector.KeyValue;
 
 /**
  * This class is used for adding data for each object in the vector layer from the userData in the