+++ src/main/java/org/opentripplanner/api/resource/GraphInspectorVectorTileResource.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.ext.vectortiles;
+package org.opentripplanner.api.resource;
 
 import static org.opentripplanner.framework.io.HttpUtils.APPLICATION_X_PROTOBUF;
 
@@ -18,13 +18,7 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
 import org.glassfish.grizzly.http.server.Request;
 import org.opentripplanner.api.model.TileJson;
-import org.opentripplanner.ext.vectortiles.layers.stations.StationsLayerBuilder;
-import org.opentripplanner.ext.vectortiles.layers.stops.StopsLayerBuilder;
-import org.opentripplanner.ext.vectortiles.layers.vehicleparkings.VehicleParkingGroupsLayerBuilder;
-import org.opentripplanner.ext.vectortiles.layers.vehicleparkings.VehicleParkingsLayerBuilder;
-import org.opentripplanner.ext.vectortiles.layers.vehiclerental.VehicleRentalPlacesLayerBuilder;
-import org.opentripplanner.ext.vectortiles.layers.vehiclerental.VehicleRentalStationsLayerBuilder;
-import org.opentripplanner.ext.vectortiles.layers.vehiclerental.VehicleRentalVehiclesLayerBuilder;
+import org.opentripplanner.inspector.vector.AreaStopsLayerBuilder;
 import org.opentripplanner.inspector.vector.LayerBuilder;
 import org.opentripplanner.inspector.vector.LayerParameters;
 import org.opentripplanner.inspector.vector.VectorTileResponseFactory;
@@ -32,23 +26,28 @@ import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.standalone.api.OtpServerRequestContext;
 import org.opentripplanner.transit.service.TransitService;
 
-@Path("/routers/{ignoreRouterId}/vectorTiles")
-public class VectorTilesResource {
+/**
+ * Slippy map vector tile API for rendering various graph information for inspection/debugging
+ * purposes.
+ */
+@Path("/routers/{ignoreRouterId}/inspector/vectortile")
+public class GraphInspectorVectorTileResource {
+
+  private static final List<LayerParameters<LayerType>> DEBUG_LAYERS = List.of(
+    new LayerParams("areaStops", LayerType.AreaStop)
+  );
 
   private final OtpServerRequestContext serverContext;
   private final String ignoreRouterId;
-  private final Locale locale;
 
-  public VectorTilesResource(
+  public GraphInspectorVectorTileResource(
     @Context OtpServerRequestContext serverContext,
-    @Context Request grizzlyRequest,
     /**
      * @deprecated The support for multiple routers are removed from OTP2.
      * See https://github.com/opentripplanner/OpenTripPlanner/issues/2760
      */
     @Deprecated @PathParam("ignoreRouterId") String ignoreRouterId
   ) {
-    this.locale = grizzlyRequest.getLocale();
     this.serverContext = serverContext;
     this.ignoreRouterId = ignoreRouterId;
   }
@@ -57,6 +56,7 @@ public class VectorTilesResource {
   @Path("/{layers}/{z}/{x}/{y}.pbf")
   @Produces(APPLICATION_X_PROTOBUF)
   public Response tileGet(
+    @Context Request grizzlyRequest,
     @PathParam("x") int x,
     @PathParam("y") int y,
     @PathParam("z") int z,
@@ -66,10 +66,10 @@ public class VectorTilesResource {
       x,
       y,
       z,
-      locale,
+      grizzlyRequest.getLocale(),
       Arrays.asList(requestedLayers.split(",")),
-      serverContext.vectorTileLayers().layers(),
-      VectorTilesResource::crateLayerBuilder,
+      DEBUG_LAYERS,
+      GraphInspectorVectorTileResource::crateLayerBuilder,
       serverContext.graph(),
       serverContext.transitService()
     );
@@ -88,7 +88,7 @@ public class VectorTilesResource {
       headers,
       requestedLayers,
       ignoreRouterId,
-      "vectorTiles",
+      "inspector/vectortile",
       serverContext.graph().getEnvelope(),
       serverContext
         .transitService()
@@ -108,42 +108,18 @@ public class VectorTilesResource {
     TransitService transitService
   ) {
     return switch (layerParameters.type()) {
-      case Stop -> new StopsLayerBuilder(transitService, layerParameters, locale);
-      case Station -> new StationsLayerBuilder(transitService, layerParameters, locale);
-      case VehicleRental -> new VehicleRentalPlacesLayerBuilder(
-        graph.getVehicleRentalService(),
-        layerParameters,
-        locale
-      );
-      case VehicleRentalStation -> new VehicleRentalStationsLayerBuilder(
-        graph.getVehicleRentalService(),
-        layerParameters,
-        locale
-      );
-      case VehicleRentalVehicle -> new VehicleRentalVehiclesLayerBuilder(
-        graph.getVehicleRentalService(),
-        layerParameters
-      );
-      case VehicleParking -> new VehicleParkingsLayerBuilder(graph, layerParameters, locale);
-      case VehicleParkingGroup -> new VehicleParkingGroupsLayerBuilder(
-        graph,
-        layerParameters,
-        locale
-      );
+      case AreaStop -> new AreaStopsLayerBuilder(transitService, layerParameters, locale);
     };
   }
 
-  public enum LayerType {
-    Stop,
-    Station,
-    VehicleRental,
-    VehicleRentalVehicle,
-    VehicleRentalStation,
-    VehicleParking,
-    VehicleParkingGroup,
+  private enum LayerType {
+    AreaStop,
   }
 
-  public interface LayersParameters<T extends Enum<T>> {
-    List<LayerParameters<T>> layers();
+  private record LayerParams(String name, LayerType type) implements LayerParameters<LayerType> {
+    @Override
+    public String mapper() {
+      return "DebugClient";
+    }
   }
 }