+++ src/main/java/org/opentripplanner/inspector/vector/KeyValue.java
@@ -1,3 +1,3 @@
-package org.opentripplanner.ext.vectortiles;
+package org.opentripplanner.inspector.vector;
 
 public record KeyValue(String key, Object value) {}